requirements.txt
Jinja2==2.11.1
num2words==0.5.10
#psycopg2>=2.8.3
pyyaml==5.3.0
rasa==1.8.1
rasa-sdk==1.8.1
numpy==1.18.1
bson==0.5.9
streamlit==0.56.0
altair==4.0.1
pandas==1.0.1
plotnine==0.6.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JWT-Extended==3.25.0
markupsafe==1.1.1
itsdangerous==1.1.0
Werkzeug==0.16.1
flask-restplus==0.13.0
nltk==3.4.5
greenlet==0.4.16
